Update to use 'Peri' from current Embassy (#2)
* Update to use 'Peri' from current Embassy

Embassy has recently added a Peri<> wrapper type around peripheral
resources. As a result the struct definitions used with
'resource_group' must include it, and the generated macros for
splitting out the resources must be able to extract the underlying
type so that they can convert it into a field name in the
'peripherals' struct.

* introduce `Peri` compatibility

* bump versions

---------

Co-authored-by: Adin Ackerman <adinackerman@gmail.com>